Hackathon2 Phase_2/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
alembic==1.13.1
sqlalchemy==2.0.23
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
celery==5.3.0
redis==5.0.1